+++ routers/web/repo/issue.go
@@ -5,58 +5,36 @@
 package repo
 
 import (
-	"bytes"
-	stdCtx "context"
 	"errors"
 	"fmt"
 	"html/template"
-	"math/big"
 	"net/http"
 	"net/url"
-	"slices"
-	"sort"
 	"strconv"
 	"strings"
 
-	activities_model "code.gitea.io/gitea/models/activities"
 	"code.gitea.io/gitea/models/db"
-	git_model "code.gitea.io/gitea/models/git"
 	issues_model "code.gitea.io/gitea/models/issues"
 	"code.gitea.io/gitea/models/organization"
 	access_model "code.gitea.io/gitea/models/perm/access"
 	project_model "code.gitea.io/gitea/models/project"
-	pull_model "code.gitea.io/gitea/models/pull"
 	repo_model "code.gitea.io/gitea/models/repo"
 	"code.gitea.io/gitea/models/unit"
 	user_model "code.gitea.io/gitea/models/user"
 	"code.gitea.io/gitea/modules/base"
-	"code.gitea.io/gitea/modules/container"
-	"code.gitea.io/gitea/modules/emoji"
-	"code.gitea.io/gitea/modules/git"
-	issue_indexer "code.gitea.io/gitea/modules/indexer/issues"
-	issue_template "code.gitea.io/gitea/modules/issue/template"
 	"code.gitea.io/gitea/modules/log"
 	"code.gitea.io/gitea/modules/markup"
 	"code.gitea.io/gitea/modules/markup/markdown"
 	"code.gitea.io/gitea/modules/optional"
-	"code.gitea.io/gitea/modules/setting"
 	api "code.gitea.io/gitea/modules/structs"
 	"code.gitea.io/gitea/modules/templates"
-	"code.gitea.io/gitea/modules/templates/vars"
 	"code.gitea.io/gitea/modules/util"
 	"code.gitea.io/gitea/modules/web"
 	"code.gitea.io/gitea/routers/common"
-	"code.gitea.io/gitea/routers/utils"
-	shared_user "code.gitea.io/gitea/routers/web/shared/user"
-	asymkey_service "code.gitea.io/gitea/services/asymkey"
 	"code.gitea.io/gitea/services/context"
-	"code.gitea.io/gitea/services/context/upload"
 	"code.gitea.io/gitea/services/convert"
 	"code.gitea.io/gitea/services/forms"
 	issue_service "code.gitea.io/gitea/services/issue"
-	pull_service "code.gitea.io/gitea/services/pull"
-	repo_service "code.gitea.io/gitea/services/repository"
-	user_service "code.gitea.io/gitea/services/user"
 )
 
 const (
@@ -140,561 +118,6 @@ func MustAllowPulls(ctx *context.Context) {
 	}
 }
 
-func issues(ctx *context.Context, milestoneID, projectID int64, isPullOption optional.Option[bool]) {
-	var err error
-	viewType := ctx.FormString("type")
-	sortType := ctx.FormString("sort")
-	types := []string{"all", "your_repositories", "assigned", "created_by", "mentioned", "review_requested", "reviewed_by"}
-	if !util.SliceContainsString(types, viewType, true) {
-		viewType = "all"
-	}
-
-	var (
-		assigneeID        = ctx.FormInt64("assignee")
-		posterID          = ctx.FormInt64("poster")
-		mentionedID       int64
-		reviewRequestedID int64
-		reviewedID        int64
-	)
-
-	if ctx.IsSigned {
-		switch viewType {
-		case "created_by":
-			posterID = ctx.Doer.ID
-		case "mentioned":
-			mentionedID = ctx.Doer.ID
-		case "assigned":
-			assigneeID = ctx.Doer.ID
-		case "review_requested":
-			reviewRequestedID = ctx.Doer.ID
-		case "reviewed_by":
-			reviewedID = ctx.Doer.ID
-		}
-	}
-
-	repo := ctx.Repo.Repository
-	var labelIDs []int64
-	// 1,-2 means including label 1 and excluding label 2
-	// 0 means issues with no label
-	// blank means labels will not be filtered for issues
-	selectLabels := ctx.FormString("labels")
-	if selectLabels == "" {
-		ctx.Data["AllLabels"] = true
-	} else if selectLabels == "0" {
-		ctx.Data["NoLabel"] = true
-	}
-	if len(selectLabels) > 0 {
-		labelIDs, err = base.StringsToInt64s(strings.Split(selectLabels, ","))
-		if err != nil {
-			ctx.Flash.Error(ctx.Tr("invalid_data", selectLabels), true)
-		}
-	}
-
-	keyword := strings.Trim(ctx.FormString("q"), " ")
-	if bytes.Contains([]byte(keyword), []byte{0x00}) {
-		keyword = ""
-	}
-
-	var mileIDs []int64
-	if milestoneID > 0 || milestoneID == db.NoConditionID { // -1 to get those issues which have no any milestone assigned
-		mileIDs = []int64{milestoneID}
-	}
-
-	var issueStats *issues_model.IssueStats
-	statsOpts := &issues_model.IssuesOptions{
-		RepoIDs:           []int64{repo.ID},
-		LabelIDs:          labelIDs,
-		MilestoneIDs:      mileIDs,
-		ProjectID:         projectID,
-		AssigneeID:        assigneeID,
-		MentionedID:       mentionedID,
-		PosterID:          posterID,
-		ReviewRequestedID: reviewRequestedID,
-		ReviewedID:        reviewedID,
-		IsPull:            isPullOption,
-		IssueIDs:          nil,
-	}
-	if keyword != "" {
-		allIssueIDs, err := issueIDsFromSearch(ctx, keyword, statsOpts)
-		if err != nil {
-			if issue_indexer.IsAvailable(ctx) {
-				ctx.ServerError("issueIDsFromSearch", err)
-				return
-			}
-			ctx.Data["IssueIndexerUnavailable"] = true
-			return
-		}
-		statsOpts.IssueIDs = allIssueIDs
-	}
-	if keyword != "" && len(statsOpts.IssueIDs) == 0 {
-		// So it did search with the keyword, but no issue found.
-		// Just set issueStats to empty.
-		issueStats = &issues_model.IssueStats{}
-	} else {
-		// So it did search with the keyword, and found some issues. It needs to get issueStats of these issues.
-		// Or the keyword is empty, so it doesn't need issueIDs as filter, just get issueStats with statsOpts.
-		issueStats, err = issues_model.GetIssueStats(ctx, statsOpts)
-		if err != nil {
-			ctx.ServerError("GetIssueStats", err)
-			return
-		}
-	}
-
-	var isShowClosed optional.Option[bool]
-	switch ctx.FormString("state") {
-	case "closed":
-		isShowClosed = optional.Some(true)
-	case "all":
-		isShowClosed = optional.None[bool]()
-	default:
-		isShowClosed = optional.Some(false)
-	}
-	// if there are closed issues and no open issues, default to showing all issues
-	if len(ctx.FormString("state")) == 0 && issueStats.OpenCount == 0 && issueStats.ClosedCount != 0 {
-		isShowClosed = optional.None[bool]()
-	}
-
-	if repo.IsTimetrackerEnabled(ctx) {
-		totalTrackedTime, err := issues_model.GetIssueTotalTrackedTime(ctx, statsOpts, isShowClosed)
-		if err != nil {
-			ctx.ServerError("GetIssueTotalTrackedTime", err)
-			return
-		}
-		ctx.Data["TotalTrackedTime"] = totalTrackedTime
-	}
-
-	archived := ctx.FormBool("archived")
-
-	page := ctx.FormInt("page")
-	if page <= 1 {
-		page = 1
-	}
-
-	var total int
-	switch {
-	case isShowClosed.Value():
-		total = int(issueStats.ClosedCount)
-	case !isShowClosed.Has():
-		total = int(issueStats.OpenCount + issueStats.ClosedCount)
-	default:
-		total = int(issueStats.OpenCount)
-	}
-	pager := context.NewPagination(total, setting.UI.IssuePagingNum, page, 5)
-
-	var issues issues_model.IssueList
-	{
-		ids, err := issueIDsFromSearch(ctx, keyword, &issues_model.IssuesOptions{
-			Paginator: &db.ListOptions{
-				Page:     pager.Paginater.Current(),
-				PageSize: setting.UI.IssuePagingNum,
-			},
-			RepoIDs:           []int64{repo.ID},
-			AssigneeID:        assigneeID,
-			PosterID:          posterID,
-			MentionedID:       mentionedID,
-			ReviewRequestedID: reviewRequestedID,
-			ReviewedID:        reviewedID,
-			MilestoneIDs:      mileIDs,
-			ProjectID:         projectID,
-			IsClosed:          isShowClosed,
-			IsPull:            isPullOption,
-			LabelIDs:          labelIDs,
-			SortType:          sortType,
-		})
-		if err != nil {
-			if issue_indexer.IsAvailable(ctx) {
-				ctx.ServerError("issueIDsFromSearch", err)
-				return
-			}
-			ctx.Data["IssueIndexerUnavailable"] = true
-			return
-		}
-		issues, err = issues_model.GetIssuesByIDs(ctx, ids, true)
-		if err != nil {
-			ctx.ServerError("GetIssuesByIDs", err)
-			return
-		}
-	}
-
-	approvalCounts, err := issues.GetApprovalCounts(ctx)
-	if err != nil {
-		ctx.ServerError("ApprovalCounts", err)
-		return
-	}
-
-	if ctx.IsSigned {
-		if err := issues.LoadIsRead(ctx, ctx.Doer.ID); err != nil {
-			ctx.ServerError("LoadIsRead", err)
-			return
-		}
-	} else {
-		for i := range issues {
-			issues[i].IsRead = true
-		}
-	}
-
-	commitStatuses, lastStatus, err := pull_service.GetIssuesAllCommitStatus(ctx, issues)
-	if err != nil {
-		ctx.ServerError("GetIssuesAllCommitStatus", err)
-		return
-	}
-	if !ctx.Repo.CanRead(unit.TypeActions) {
-		for key := range commitStatuses {
-			git_model.CommitStatusesHideActionsURL(ctx, commitStatuses[key])
-		}
-	}
-
-	if err := issues.LoadAttributes(ctx); err != nil {
-		ctx.ServerError("issues.LoadAttributes", err)
-		return
-	}
-
-	ctx.Data["Issues"] = issues
-	ctx.Data["CommitLastStatus"] = lastStatus
-	ctx.Data["CommitStatuses"] = commitStatuses
-
-	// Get assignees.
-	assigneeUsers, err := repo_model.GetRepoAssignees(ctx, repo)
-	if err != nil {
-		ctx.ServerError("GetRepoAssignees", err)
-		return
-	}
-	ctx.Data["Assignees"] = shared_user.MakeSelfOnTop(ctx.Doer, assigneeUsers)
-
-	handleTeamMentions(ctx)
-	if ctx.Written() {
-		return
-	}
-
-	labels, err := issues_model.GetLabelsByRepoID(ctx, repo.ID, "", db.ListOptions{})
-	if err != nil {
-		ctx.ServerError("GetLabelsByRepoID", err)
-		return
-	}
-
-	if repo.Owner.IsOrganization() {
-		orgLabels, err := issues_model.GetLabelsByOrgID(ctx, repo.Owner.ID, ctx.FormString("sort"), db.ListOptions{})
-		if err != nil {
-			ctx.ServerError("GetLabelsByOrgID", err)
-			return
-		}
-
-		ctx.Data["OrgLabels"] = orgLabels
-		labels = append(labels, orgLabels...)
-	}
-
-	// Get the exclusive scope for every label ID
-	labelExclusiveScopes := make([]string, 0, len(labelIDs))
-	for _, labelID := range labelIDs {
-		foundExclusiveScope := false
-		for _, label := range labels {
-			if label.ID == labelID || label.ID == -labelID {
-				labelExclusiveScopes = append(labelExclusiveScopes, label.ExclusiveScope())
-				foundExclusiveScope = true
-				break
-			}
-		}
-		if !foundExclusiveScope {
-			labelExclusiveScopes = append(labelExclusiveScopes, "")
-		}
-	}
-
-	for _, l := range labels {
-		l.LoadSelectedLabelsAfterClick(labelIDs, labelExclusiveScopes)
-	}
-	ctx.Data["Labels"] = labels
-	ctx.Data["NumLabels"] = len(labels)
-
-	if ctx.FormInt64("assignee") == 0 {
-		assigneeID = 0 // Reset ID to prevent unexpected selection of assignee.
-	}
-
-	ctx.Data["IssueRefEndNames"], ctx.Data["IssueRefURLs"] = issue_service.GetRefEndNamesAndURLs(issues, ctx.Repo.RepoLink)
-
-	ctx.Data["ApprovalCounts"] = func(issueID int64, typ string) int64 {
-		counts, ok := approvalCounts[issueID]
-		if !ok || len(counts) == 0 {
-			return 0
-		}
-		reviewTyp := issues_model.ReviewTypeApprove
-		if typ == "reject" {
-			reviewTyp = issues_model.ReviewTypeReject
-		} else if typ == "waiting" {
-			reviewTyp = issues_model.ReviewTypeRequest
-		}
-		for _, count := range counts {
-			if count.Type == reviewTyp {
-				return count.Count
-			}
-		}
-		return 0
-	}
-
-	retrieveProjectsForIssueList(ctx, repo)
-	if ctx.Written() {
-		return
-	}
-
-	pinned, err := issues_model.GetPinnedIssues(ctx, repo.ID, isPullOption.Value())
-	if err != nil {
-		ctx.ServerError("GetPinnedIssues", err)
-		return
-	}
-
-	ctx.Data["PinnedIssues"] = pinned
-	ctx.Data["IsRepoAdmin"] = ctx.IsSigned && (ctx.Repo.IsAdmin() || ctx.Doer.IsAdmin)
-	ctx.Data["IssueStats"] = issueStats
-	ctx.Data["OpenCount"] = issueStats.OpenCount
-	ctx.Data["ClosedCount"] = issueStats.ClosedCount
-	linkStr := "%s?q=%s&type=%s&sort=%s&state=%s&labels=%s&milestone=%d&project=%d&assignee=%d&poster=%d&archived=%t"
-	ctx.Data["AllStatesLink"] = fmt.Sprintf(linkStr, ctx.Link,
-		url.QueryEscape(keyword), url.QueryEscape(viewType), url.QueryEscape(sortType), "all", url.QueryEscape(selectLabels),
-		milestoneID, projectID, assigneeID, posterID, archived)
-	ctx.Data["OpenLink"] = fmt.Sprintf(linkStr, ctx.Link,
-		url.QueryEscape(keyword), url.QueryEscape(viewType), url.QueryEscape(sortType), "open", url.QueryEscape(selectLabels),
-		milestoneID, projectID, assigneeID, posterID, archived)
-	ctx.Data["ClosedLink"] = fmt.Sprintf(linkStr, ctx.Link,
-		url.QueryEscape(keyword), url.QueryEscape(viewType), url.QueryEscape(sortType), "closed", url.QueryEscape(selectLabels),
-		milestoneID, projectID, assigneeID, posterID, archived)
-	ctx.Data["SelLabelIDs"] = labelIDs
-	ctx.Data["SelectLabels"] = selectLabels
-	ctx.Data["ViewType"] = viewType
-	ctx.Data["SortType"] = sortType
-	ctx.Data["MilestoneID"] = milestoneID
-	ctx.Data["ProjectID"] = projectID
-	ctx.Data["AssigneeID"] = assigneeID
-	ctx.Data["PosterID"] = posterID
-	ctx.Data["Keyword"] = keyword
-	ctx.Data["IsShowClosed"] = isShowClosed
-	switch {
-	case isShowClosed.Value():
-		ctx.Data["State"] = "closed"
-	case !isShowClosed.Has():
-		ctx.Data["State"] = "all"
-	default:
-		ctx.Data["State"] = "open"
-	}
-	ctx.Data["ShowArchivedLabels"] = archived
-
-	pager.AddParamString("q", keyword)
-	pager.AddParamString("type", viewType)
-	pager.AddParamString("sort", sortType)
-	pager.AddParamString("state", fmt.Sprint(ctx.Data["State"]))
-	pager.AddParamString("labels", fmt.Sprint(selectLabels))
-	pager.AddParamString("milestone", fmt.Sprint(milestoneID))
-	pager.AddParamString("project", fmt.Sprint(projectID))
-	pager.AddParamString("assignee", fmt.Sprint(assigneeID))
-	pager.AddParamString("poster", fmt.Sprint(posterID))
-	pager.AddParamString("archived", fmt.Sprint(archived))
-
-	ctx.Data["Page"] = pager
-}
-
-func issueIDsFromSearch(ctx *context.Context, keyword string, opts *issues_model.IssuesOptions) ([]int64, error) {
-	ids, _, err := issue_indexer.SearchIssues(ctx, issue_indexer.ToSearchOptions(keyword, opts))
-	if err != nil {
-		return nil, fmt.Errorf("SearchIssues: %w", err)
-	}
-	return ids, nil
-}
-
-// Issues render issues page
-func Issues(ctx *context.Context) {
-	isPullList := ctx.PathParam(":type") == "pulls"
-	if isPullList {
-		MustAllowPulls(ctx)
-		if ctx.Written() {
-			return
-		}
-		ctx.Data["Title"] = ctx.Tr("repo.pulls")
-		ctx.Data["PageIsPullList"] = true
-	} else {
-		MustEnableIssues(ctx)
-		if ctx.Written() {
-			return
-		}
-		ctx.Data["Title"] = ctx.Tr("repo.issues")
-		ctx.Data["PageIsIssueList"] = true
-		ctx.Data["NewIssueChooseTemplate"] = issue_service.HasTemplatesOrContactLinks(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	}
-
-	issues(ctx, ctx.FormInt64("milestone"), ctx.FormInt64("project"), optional.Some(isPullList))
-	if ctx.Written() {
-		return
-	}
-
-	renderMilestones(ctx)
-	if ctx.Written() {
-		return
-	}
-
-	ctx.Data["CanWriteIssuesOrPulls"] = ctx.Repo.CanWriteIssuesOrPulls(isPullList)
-
-	ctx.HTML(http.StatusOK, tplIssues)
-}
-
-func renderMilestones(ctx *context.Context) {
-	// Get milestones
-	milestones, err := db.Find[issues_model.Milestone](ctx, issues_model.FindMilestoneOptions{
-		RepoID: ctx.Repo.Repository.ID,
-	})
-	if err != nil {
-		ctx.ServerError("GetAllRepoMilestones", err)
-		return
-	}
-
-	openMilestones, closedMilestones := issues_model.MilestoneList{}, issues_model.MilestoneList{}
-	for _, milestone := range milestones {
-		if milestone.IsClosed {
-			closedMilestones = append(closedMilestones, milestone)
-		} else {
-			openMilestones = append(openMilestones, milestone)
-		}
-	}
-	ctx.Data["OpenMilestones"] = openMilestones
-	ctx.Data["ClosedMilestones"] = closedMilestones
-}
-
-type issueSidebarMilestoneData struct {
-	SelectedMilestoneID int64
-	OpenMilestones      []*issues_model.Milestone
-	ClosedMilestones    []*issues_model.Milestone
-}
-
-type issueSidebarAssigneesData struct {
-	SelectedAssigneeIDs string
-	CandidateAssignees  []*user_model.User
-}
-
-type IssuePageMetaData struct {
-	RepoLink             string
-	Repository           *repo_model.Repository
-	Issue                *issues_model.Issue
-	IsPullRequest        bool
-	CanModifyIssueOrPull bool
-
-	ReviewersData  *issueSidebarReviewersData
-	LabelsData     *issueSidebarLabelsData
-	MilestonesData *issueSidebarMilestoneData
-	ProjectsData   *issueSidebarProjectsData
-	AssigneesData  *issueSidebarAssigneesData
-}
-
-func retrieveRepoIssueMetaData(ctx *context.Context, repo *repo_model.Repository, issue *issues_model.Issue, isPull bool) *IssuePageMetaData {
-	data := &IssuePageMetaData{
-		RepoLink:      ctx.Repo.RepoLink,
-		Repository:    repo,
-		Issue:         issue,
-		IsPullRequest: isPull,
-
-		ReviewersData:  &issueSidebarReviewersData{},
-		LabelsData:     &issueSidebarLabelsData{},
-		MilestonesData: &issueSidebarMilestoneData{},
-		ProjectsData:   &issueSidebarProjectsData{},
-		AssigneesData:  &issueSidebarAssigneesData{},
-	}
-	ctx.Data["IssuePageMetaData"] = data
-
-	if isPull {
-		data.retrieveReviewersData(ctx)
-		if ctx.Written() {
-			return data
-		}
-	}
-	data.retrieveLabelsData(ctx)
-	if ctx.Written() {
-		return data
-	}
-
-	data.CanModifyIssueOrPull = ctx.Repo.CanWriteIssuesOrPulls(isPull) && !ctx.Repo.Repository.IsArchived
-	if !data.CanModifyIssueOrPull {
-		return data
-	}
-
-	data.retrieveAssigneesDataForIssueWriter(ctx)
-	if ctx.Written() {
-		return data
-	}
-
-	data.retrieveMilestonesDataForIssueWriter(ctx)
-	if ctx.Written() {
-		return data
-	}
-
-	data.retrieveProjectsDataForIssueWriter(ctx)
-	if ctx.Written() {
-		return data
-	}
-
-	PrepareBranchList(ctx)
-	if ctx.Written() {
-		return data
-	}
-
-	ctx.Data["CanCreateIssueDependencies"] = ctx.Repo.CanCreateIssueDependencies(ctx, ctx.Doer, isPull)
-	return data
-}
-
-func (d *IssuePageMetaData) retrieveMilestonesDataForIssueWriter(ctx *context.Context) {
-	var err error
-	if d.Issue != nil {
-		d.MilestonesData.SelectedMilestoneID = d.Issue.MilestoneID
-	}
-	d.MilestonesData.OpenMilestones, err = db.Find[issues_model.Milestone](ctx, issues_model.FindMilestoneOptions{
-		RepoID:   d.Repository.ID,
-		IsClosed: optional.Some(false),
-	})
-	if err != nil {
-		ctx.ServerError("GetMilestones", err)
-		return
-	}
-	d.MilestonesData.ClosedMilestones, err = db.Find[issues_model.Milestone](ctx, issues_model.FindMilestoneOptions{
-		RepoID:   d.Repository.ID,
-		IsClosed: optional.Some(true),
-	})
-	if err != nil {
-		ctx.ServerError("GetMilestones", err)
-		return
-	}
-}
-
-func (d *IssuePageMetaData) retrieveAssigneesDataForIssueWriter(ctx *context.Context) {
-	var err error
-	d.AssigneesData.CandidateAssignees, err = repo_model.GetRepoAssignees(ctx, d.Repository)
-	if err != nil {
-		ctx.ServerError("GetRepoAssignees", err)
-		return
-	}
-	d.AssigneesData.CandidateAssignees = shared_user.MakeSelfOnTop(ctx.Doer, d.AssigneesData.CandidateAssignees)
-	if d.Issue != nil {
-		_ = d.Issue.LoadAssignees(ctx)
-		ids := make([]string, 0, len(d.Issue.Assignees))
-		for _, a := range d.Issue.Assignees {
-			ids = append(ids, strconv.FormatInt(a.ID, 10))
-		}
-		d.AssigneesData.SelectedAssigneeIDs = strings.Join(ids, ",")
-	}
-	// FIXME: this is a tricky part which writes ctx.Data["Mentionable*"]
-	handleTeamMentions(ctx)
-}
-
-func retrieveProjectsForIssueList(ctx *context.Context, repo *repo_model.Repository) {
-	ctx.Data["OpenProjects"], ctx.Data["ClosedProjects"] = retrieveProjectsInternal(ctx, repo)
-}
-
-type issueSidebarProjectsData struct {
-	SelectedProjectID int64
-	OpenProjects      []*project_model.Project
-	ClosedProjects    []*project_model.Project
-}
-
-func (d *IssuePageMetaData) retrieveProjectsDataForIssueWriter(ctx *context.Context) {
-	if d.Issue != nil && d.Issue.Project != nil {
-		d.ProjectsData.SelectedProjectID = d.Issue.Project.ID
-	}
-	d.ProjectsData.OpenProjects, d.ProjectsData.ClosedProjects = retrieveProjectsInternal(ctx, ctx.Repo.Repository)
-}
-
 func retrieveProjectsInternal(ctx *context.Context, repo *repo_model.Repository) (open, closed []*project_model.Project) {
 	// Distinguish whether the owner of the repository
 	// is an individual or an organization
@@ -757,1432 +180,6 @@ func retrieveProjectsInternal(ctx *context.Context, repo *repo_model.Repository)
 		closedProjects = append(closedProjects, closedProjects2...)
 	}
 	return openProjects, closedProjects
-}
-
-// repoReviewerSelection items to bee shown
-type repoReviewerSelection struct {
-	IsTeam         bool
-	Team           *organization.Team
-	User           *user_model.User
-	Review         *issues_model.Review
-	CanBeDismissed bool
-	CanChange      bool
-	Requested      bool
-	ItemID         int64
-}
-
-type issueSidebarReviewersData struct {
-	CanChooseReviewer    bool
-	OriginalReviews      issues_model.ReviewList
-	TeamReviewers        []*repoReviewerSelection
-	Reviewers            []*repoReviewerSelection
-	CurrentPullReviewers []*repoReviewerSelection
-}
-
-// RetrieveRepoReviewers find all reviewers of a repository. If issue is nil, it means the doer is creating a new PR.
-func (d *IssuePageMetaData) retrieveReviewersData(ctx *context.Context) {
-	data := d.ReviewersData
-	repo := d.Repository
-	if ctx.Doer != nil && ctx.IsSigned {
-		if d.Issue == nil {
-			data.CanChooseReviewer = true
-		} else {
-			data.CanChooseReviewer = issue_service.CanDoerChangeReviewRequests(ctx, ctx.Doer, repo, d.Issue)
-		}
-	}
-
-	var posterID int64
-	var isClosed bool
-	var reviews issues_model.ReviewList
-
-	if d.Issue == nil {
-		posterID = ctx.Doer.ID
-	} else {
-		posterID = d.Issue.PosterID
-		if d.Issue.OriginalAuthorID > 0 {
-			posterID = 0 // for migrated PRs, no poster ID
-		}
-
-		isClosed = d.Issue.IsClosed || d.Issue.PullRequest.HasMerged
-
-		originalAuthorReviews, err := issues_model.GetReviewersFromOriginalAuthorsByIssueID(ctx, d.Issue.ID)
-		if err != nil {
-			ctx.ServerError("GetReviewersFromOriginalAuthorsByIssueID", err)
-			return
-		}
-		data.OriginalReviews = originalAuthorReviews
-
-		reviews, err = issues_model.GetReviewsByIssueID(ctx, d.Issue.ID)
-		if err != nil {
-			ctx.ServerError("GetReviewersByIssueID", err)
-			return
-		}
-		if len(reviews) == 0 && !data.CanChooseReviewer {
-			return
-		}
-	}
-
-	var (
-		pullReviews         []*repoReviewerSelection
-		reviewersResult     []*repoReviewerSelection
-		teamReviewersResult []*repoReviewerSelection
-		teamReviewers       []*organization.Team
-		reviewers           []*user_model.User
-	)
-
-	if data.CanChooseReviewer {
-		var err error
-		reviewers, err = repo_model.GetReviewers(ctx, repo, ctx.Doer.ID, posterID)
-		if err != nil {
-			ctx.ServerError("GetReviewers", err)
-			return
-		}
-
-		teamReviewers, err = repo_service.GetReviewerTeams(ctx, repo)
-		if err != nil {
-			ctx.ServerError("GetReviewerTeams", err)
-			return
-		}
-
-		if len(reviewers) > 0 {
-			reviewersResult = make([]*repoReviewerSelection, 0, len(reviewers))
-		}
-
-		if len(teamReviewers) > 0 {
-			teamReviewersResult = make([]*repoReviewerSelection, 0, len(teamReviewers))
-		}
-	}
-
-	pullReviews = make([]*repoReviewerSelection, 0, len(reviews))
-
-	for _, review := range reviews {
-		tmp := &repoReviewerSelection{
-			Requested: review.Type == issues_model.ReviewTypeRequest,
-			Review:    review,
-			ItemID:    review.ReviewerID,
-		}
-		if review.ReviewerTeamID > 0 {
-			tmp.IsTeam = true
-			tmp.ItemID = -review.ReviewerTeamID
-		}
-
-		if data.CanChooseReviewer {
-			// Users who can choose reviewers can also remove review requests
-			tmp.CanChange = true
-		} else if ctx.Doer != nil && ctx.Doer.ID == review.ReviewerID && review.Type == issues_model.ReviewTypeRequest {
-			// A user can refuse review requests
-			tmp.CanChange = true
-		}
-
-		pullReviews = append(pullReviews, tmp)
-
-		if data.CanChooseReviewer {
-			if tmp.IsTeam {
-				teamReviewersResult = append(teamReviewersResult, tmp)
-			} else {
-				reviewersResult = append(reviewersResult, tmp)
-			}
-		}
-	}
-
-	if len(pullReviews) > 0 {
-		// Drop all non-existing users and teams from the reviews
-		currentPullReviewers := make([]*repoReviewerSelection, 0, len(pullReviews))
-		for _, item := range pullReviews {
-			if item.Review.ReviewerID > 0 {
-				if err := item.Review.LoadReviewer(ctx); err != nil {
-					if user_model.IsErrUserNotExist(err) {
-						continue
-					}
-					ctx.ServerError("LoadReviewer", err)
-					return
-				}
-				item.User = item.Review.Reviewer
-			} else if item.Review.ReviewerTeamID > 0 {
-				if err := item.Review.LoadReviewerTeam(ctx); err != nil {
-					if organization.IsErrTeamNotExist(err) {
-						continue
-					}
-					ctx.ServerError("LoadReviewerTeam", err)
-					return
-				}
-				item.Team = item.Review.ReviewerTeam
-			} else {
-				continue
-			}
-			item.CanBeDismissed = ctx.Repo.Permission.IsAdmin() && !isClosed &&
-				(item.Review.Type == issues_model.ReviewTypeApprove || item.Review.Type == issues_model.ReviewTypeReject)
-			currentPullReviewers = append(currentPullReviewers, item)
-		}
-		data.CurrentPullReviewers = currentPullReviewers
-	}
-
-	if data.CanChooseReviewer && reviewersResult != nil {
-		preadded := len(reviewersResult)
-		for _, reviewer := range reviewers {
-			found := false
-		reviewAddLoop:
-			for _, tmp := range reviewersResult[:preadded] {
-				if tmp.ItemID == reviewer.ID {
-					tmp.User = reviewer
-					found = true
-					break reviewAddLoop
-				}
-			}
-
-			if found {
-				continue
-			}
-
-			reviewersResult = append(reviewersResult, &repoReviewerSelection{
-				IsTeam:    false,
-				CanChange: true,
-				User:      reviewer,
-				ItemID:    reviewer.ID,
-			})
-		}
-
-		data.Reviewers = reviewersResult
-	}
-
-	if data.CanChooseReviewer && teamReviewersResult != nil {
-		preadded := len(teamReviewersResult)
-		for _, team := range teamReviewers {
-			found := false
-		teamReviewAddLoop:
-			for _, tmp := range teamReviewersResult[:preadded] {
-				if tmp.ItemID == -team.ID {
-					tmp.Team = team
-					found = true
-					break teamReviewAddLoop
-				}
-			}
-
-			if found {
-				continue
-			}
-
-			teamReviewersResult = append(teamReviewersResult, &repoReviewerSelection{
-				IsTeam:    true,
-				CanChange: true,
-				Team:      team,
-				ItemID:    -team.ID,
-			})
-		}
-
-		data.TeamReviewers = teamReviewersResult
-	}
-}
-
-type issueSidebarLabelsData struct {
-	AllLabels        []*issues_model.Label
-	RepoLabels       []*issues_model.Label
-	OrgLabels        []*issues_model.Label
-	SelectedLabelIDs string
-}
-
-func makeSelectedStringIDs[KeyType, ItemType comparable](
-	allLabels []*issues_model.Label, candidateKey func(candidate *issues_model.Label) KeyType,
-	selectedItems []ItemType, selectedKey func(selected ItemType) KeyType,
-) string {
-	selectedIDSet := make(container.Set[string])
-	allLabelMap := map[KeyType]*issues_model.Label{}
-	for _, label := range allLabels {
-		allLabelMap[candidateKey(label)] = label
-	}
-	for _, item := range selectedItems {
-		if label, ok := allLabelMap[selectedKey(item)]; ok {
-			label.IsChecked = true
-			selectedIDSet.Add(strconv.FormatInt(label.ID, 10))
-		}
-	}
-	ids := selectedIDSet.Values()
-	sort.Strings(ids)
-	return strings.Join(ids, ",")
-}
-
-func (d *issueSidebarLabelsData) SetSelectedLabels(labels []*issues_model.Label) {
-	d.SelectedLabelIDs = makeSelectedStringIDs(
-		d.AllLabels, func(label *issues_model.Label) int64 { return label.ID },
-		labels, func(label *issues_model.Label) int64 { return label.ID },
-	)
-}
-
-func (d *issueSidebarLabelsData) SetSelectedLabelNames(labelNames []string) {
-	d.SelectedLabelIDs = makeSelectedStringIDs(
-		d.AllLabels, func(label *issues_model.Label) string { return strings.ToLower(label.Name) },
-		labelNames, strings.ToLower,
-	)
-}
-
-func (d *issueSidebarLabelsData) SetSelectedLabelIDs(labelIDs []int64) {
-	d.SelectedLabelIDs = makeSelectedStringIDs(
-		d.AllLabels, func(label *issues_model.Label) int64 { return label.ID },
-		labelIDs, func(labelID int64) int64 { return labelID },
-	)
-}
-
-func (d *IssuePageMetaData) retrieveLabelsData(ctx *context.Context) {
-	repo := d.Repository
-	labelsData := d.LabelsData
-
-	labels, err := issues_model.GetLabelsByRepoID(ctx, repo.ID, "", db.ListOptions{})
-	if err != nil {
-		ctx.ServerError("GetLabelsByRepoID", err)
-		return
-	}
-	labelsData.RepoLabels = labels
-
-	if repo.Owner.IsOrganization() {
-		orgLabels, err := issues_model.GetLabelsByOrgID(ctx, repo.Owner.ID, ctx.FormString("sort"), db.ListOptions{})
-		if err != nil {
-			return
-		}
-		labelsData.OrgLabels = orgLabels
-	}
-	labelsData.AllLabels = append(labelsData.AllLabels, labelsData.RepoLabels...)
-	labelsData.AllLabels = append(labelsData.AllLabels, labelsData.OrgLabels...)
-}
-
-// Tries to load and set an issue template. The first return value indicates if a template was loaded.
-func setTemplateIfExists(ctx *context.Context, ctxDataKey string, possibleFiles []string, metaData *IssuePageMetaData) (bool, map[string]error) {
-	commit, err := ctx.Repo.GitRepo.GetBranchCommit(ctx.Repo.Repository.DefaultBranch)
-	if err != nil {
-		return false, nil
-	}
-
-	templateCandidates := make([]string, 0, 1+len(possibleFiles))
-	if t := ctx.FormString("template"); t != "" {
-		templateCandidates = append(templateCandidates, t)
-	}
-	templateCandidates = append(templateCandidates, possibleFiles...) // Append files to the end because they should be fallback
-
-	templateErrs := map[string]error{}
-	for _, filename := range templateCandidates {
-		if ok, _ := commit.HasFile(filename); !ok {
-			continue
-		}
-		template, err := issue_template.UnmarshalFromCommit(commit, filename)
-		if err != nil {
-			templateErrs[filename] = err
-			continue
-		}
-		ctx.Data[issueTemplateTitleKey] = template.Title
-		ctx.Data[ctxDataKey] = template.Content
-
-		if template.Type() == api.IssueTemplateTypeYaml {
-			// Replace field default values by values from query
-			for _, field := range template.Fields {
-				fieldValue := ctx.FormString("field:" + field.ID)
-				if fieldValue != "" {
-					field.Attributes["value"] = fieldValue
-				}
-			}
-
-			ctx.Data["Fields"] = template.Fields
-			ctx.Data["TemplateFile"] = template.FileName
-		}
-
-		metaData.LabelsData.SetSelectedLabelNames(template.Labels)
-
-		selectedAssigneeIDStrings := make([]string, 0, len(template.Assignees))
-		if userIDs, err := user_model.GetUserIDsByNames(ctx, template.Assignees, true); err == nil {
-			for _, userID := range userIDs {
-				selectedAssigneeIDStrings = append(selectedAssigneeIDStrings, strconv.FormatInt(userID, 10))
-			}
-		}
-		metaData.AssigneesData.SelectedAssigneeIDs = strings.Join(selectedAssigneeIDStrings, ",")
-
-		if template.Ref != "" && !strings.HasPrefix(template.Ref, "refs/") { // Assume that the ref intended is always a branch - for tags users should use refs/tags/<ref>
-			template.Ref = git.BranchPrefix + template.Ref
-		}
-
-		ctx.Data["Reference"] = template.Ref
-		ctx.Data["RefEndName"] = git.RefName(template.Ref).ShortName()
-		return true, templateErrs
-	}
-	return false, templateErrs
-}
-
-// NewIssue render creating issue page
-func NewIssue(ctx *context.Context) {
-	issueConfig, _ := issue_service.GetTemplateConfigFromDefaultBranch(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	hasTemplates := issue_service.HasTemplatesOrContactLinks(ctx.Repo.Repository, ctx.Repo.GitRepo)
-
-	ctx.Data["Title"] = ctx.Tr("repo.issues.new")
-	ctx.Data["PageIsIssueList"] = true
-	ctx.Data["NewIssueChooseTemplate"] = hasTemplates
-	ctx.Data["PullRequestWorkInProgressPrefixes"] = setting.Repository.PullRequest.WorkInProgressPrefixes
-	title := ctx.FormString("title")
-	ctx.Data["TitleQuery"] = title
-	body := ctx.FormString("body")
-	ctx.Data["BodyQuery"] = body
-
-	isProjectsEnabled := ctx.Repo.CanRead(unit.TypeProjects)
-	ctx.Data["IsProjectsEnabled"] = isProjectsEnabled
-	ctx.Data["IsAttachmentEnabled"] = setting.Attachment.Enabled
-	upload.AddUploadContext(ctx, "comment")
-
-	pageMetaData := retrieveRepoIssueMetaData(ctx, ctx.Repo.Repository, nil, false)
-	if ctx.Written() {
-		return
-	}
-
-	pageMetaData.MilestonesData.SelectedMilestoneID = ctx.FormInt64("milestone")
-	pageMetaData.ProjectsData.SelectedProjectID = ctx.FormInt64("project")
-	if pageMetaData.ProjectsData.SelectedProjectID > 0 {
-		if len(ctx.Req.URL.Query().Get("project")) > 0 {
-			ctx.Data["redirect_after_creation"] = "project"
-		}
-	}
-
-	tags, err := repo_model.GetTagNamesByRepoID(ctx, ctx.Repo.Repository.ID)
-	if err != nil {
-		ctx.ServerError("GetTagNamesByRepoID", err)
-		return
-	}
-	ctx.Data["Tags"] = tags
-
-	ret := issue_service.ParseTemplatesFromDefaultBranch(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	templateLoaded, errs := setTemplateIfExists(ctx, issueTemplateKey, IssueTemplateCandidates, pageMetaData)
-	for k, v := range errs {
-		ret.TemplateErrors[k] = v
-	}
-	if ctx.Written() {
-		return
-	}
-
-	if len(ret.TemplateErrors) > 0 {
-		ctx.Flash.Warning(renderErrorOfTemplates(ctx, ret.TemplateErrors), true)
-	}
-
-	ctx.Data["HasIssuesOrPullsWritePermission"] = ctx.Repo.CanWrite(unit.TypeIssues)
-
-	if !issueConfig.BlankIssuesEnabled && hasTemplates && !templateLoaded {
-		// The "issues/new" and "issues/new/choose" share the same query parameters "project" and "milestone", if blank issues are disabled, just redirect to the "issues/choose" page with these parameters.
-		ctx.Redirect(fmt.Sprintf("%s/issues/new/choose?%s", ctx.Repo.Repository.Link(), ctx.Req.URL.RawQuery), http.StatusSeeOther)
-		return
-	}
-
-	ctx.HTML(http.StatusOK, tplIssueNew)
-}
-
-func renderErrorOfTemplates(ctx *context.Context, errs map[string]error) template.HTML {
-	var files []string
-	for k := range errs {
-		files = append(files, k)
-	}
-	sort.Strings(files) // keep the output stable
-
-	var lines []string
-	for _, file := range files {
-		lines = append(lines, fmt.Sprintf("%s: %v", file, errs[file]))
-	}
-
-	flashError, err := ctx.RenderToHTML(tplAlertDetails, map[string]any{
-		"Message": ctx.Tr("repo.issues.choose.ignore_invalid_templates"),
-		"Summary": ctx.Tr("repo.issues.choose.invalid_templates", len(errs)),
-		"Details": utils.SanitizeFlashErrorString(strings.Join(lines, "\n")),
-	})
-	if err != nil {
-		log.Debug("render flash error: %v", err)
-		flashError = ctx.Locale.Tr("repo.issues.choose.ignore_invalid_templates")
-	}
-	return flashError
-}
-
-// NewIssueChooseTemplate render creating issue from template page
-func NewIssueChooseTemplate(ctx *context.Context) {
-	ctx.Data["Title"] = ctx.Tr("repo.issues.new")
-	ctx.Data["PageIsIssueList"] = true
-
-	ret := issue_service.ParseTemplatesFromDefaultBranch(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	ctx.Data["IssueTemplates"] = ret.IssueTemplates
-
-	if len(ret.TemplateErrors) > 0 {
-		ctx.Flash.Warning(renderErrorOfTemplates(ctx, ret.TemplateErrors), true)
-	}
-
-	if !issue_service.HasTemplatesOrContactLinks(ctx.Repo.Repository, ctx.Repo.GitRepo) {
-		// The "issues/new" and "issues/new/choose" share the same query parameters "project" and "milestone", if no template here, just redirect to the "issues/new" page with these parameters.
-		ctx.Redirect(fmt.Sprintf("%s/issues/new?%s", ctx.Repo.Repository.Link(), ctx.Req.URL.RawQuery), http.StatusSeeOther)
-		return
-	}
-
-	issueConfig, err := issue_service.GetTemplateConfigFromDefaultBranch(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	ctx.Data["IssueConfig"] = issueConfig
-	ctx.Data["IssueConfigError"] = err // ctx.Flash.Err makes problems here
-
-	ctx.Data["milestone"] = ctx.FormInt64("milestone")
-	ctx.Data["project"] = ctx.FormInt64("project")
-
-	ctx.HTML(http.StatusOK, tplIssueChoose)
-}
-
-// DeleteIssue deletes an issue
-func DeleteIssue(ctx *context.Context) {
-	issue := GetActionIssue(ctx)
-	if ctx.Written() {
-		return
-	}
-
-	if err := issue_service.DeleteIssue(ctx, ctx.Doer, ctx.Repo.GitRepo, issue); err != nil {
-		ctx.ServerError("DeleteIssueByID", err)
-		return
-	}
-
-	if issue.IsPull {
-		ctx.Redirect(fmt.Sprintf("%s/pulls", ctx.Repo.Repository.Link()), http.StatusSeeOther)
-		return
-	}
-
-	ctx.Redirect(fmt.Sprintf("%s/issues", ctx.Repo.Repository.Link()), http.StatusSeeOther)
-}
-
-func toSet[ItemType any, KeyType comparable](slice []ItemType, keyFunc func(ItemType) KeyType) container.Set[KeyType] {
-	s := make(container.Set[KeyType])
-	for _, item := range slice {
-		s.Add(keyFunc(item))
-	}
-	return s
-}
-
-// ValidateRepoMetasForNewIssue check and returns repository's meta information
-func ValidateRepoMetasForNewIssue(ctx *context.Context, form forms.CreateIssueForm, isPull bool) (ret struct {
-	LabelIDs, AssigneeIDs  []int64
-	MilestoneID, ProjectID int64
-
-	Reviewers     []*user_model.User
-	TeamReviewers []*organization.Team
-},
-) {
-	pageMetaData := retrieveRepoIssueMetaData(ctx, ctx.Repo.Repository, nil, isPull)
-	if ctx.Written() {
-		return ret
-	}
-
-	inputLabelIDs, _ := base.StringsToInt64s(strings.Split(form.LabelIDs, ","))
-	candidateLabels := toSet(pageMetaData.LabelsData.AllLabels, func(label *issues_model.Label) int64 { return label.ID })
-	if len(inputLabelIDs) > 0 && !candidateLabels.Contains(inputLabelIDs...) {
-		ctx.NotFound("", nil)
-		return ret
-	}
-	pageMetaData.LabelsData.SetSelectedLabelIDs(inputLabelIDs)
-
-	allMilestones := append(slices.Clone(pageMetaData.MilestonesData.OpenMilestones), pageMetaData.MilestonesData.ClosedMilestones...)
-	candidateMilestones := toSet(allMilestones, func(milestone *issues_model.Milestone) int64 { return milestone.ID })
-	if form.MilestoneID > 0 && !candidateMilestones.Contains(form.MilestoneID) {
-		ctx.NotFound("", nil)
-		return ret
-	}
-	pageMetaData.MilestonesData.SelectedMilestoneID = form.MilestoneID
-
-	allProjects := append(slices.Clone(pageMetaData.ProjectsData.OpenProjects), pageMetaData.ProjectsData.ClosedProjects...)
-	candidateProjects := toSet(allProjects, func(project *project_model.Project) int64 { return project.ID })
-	if form.ProjectID > 0 && !candidateProjects.Contains(form.ProjectID) {
-		ctx.NotFound("", nil)
-		return ret
-	}
-	pageMetaData.ProjectsData.SelectedProjectID = form.ProjectID
-
-	candidateAssignees := toSet(pageMetaData.AssigneesData.CandidateAssignees, func(user *user_model.User) int64 { return user.ID })
-	inputAssigneeIDs, _ := base.StringsToInt64s(strings.Split(form.AssigneeIDs, ","))
-	if len(inputAssigneeIDs) > 0 && !candidateAssignees.Contains(inputAssigneeIDs...) {
-		ctx.NotFound("", nil)
-		return ret
-	}
-	pageMetaData.AssigneesData.SelectedAssigneeIDs = form.AssigneeIDs
-
-	// Check if the passed reviewers (user/team) actually exist
-	var reviewers []*user_model.User
-	var teamReviewers []*organization.Team
-	reviewerIDs, _ := base.StringsToInt64s(strings.Split(form.ReviewerIDs, ","))
-	if isPull && len(reviewerIDs) > 0 {
-		userReviewersMap := map[int64]*user_model.User{}
-		teamReviewersMap := map[int64]*organization.Team{}
-		for _, r := range pageMetaData.ReviewersData.Reviewers {
-			userReviewersMap[r.User.ID] = r.User
-		}
-		for _, r := range pageMetaData.ReviewersData.TeamReviewers {
-			teamReviewersMap[r.Team.ID] = r.Team
-		}
-		for _, rID := range reviewerIDs {
-			if rID < 0 { // negative reviewIDs represent team requests
-				team, ok := teamReviewersMap[-rID]
-				if !ok {
-					ctx.NotFound("", nil)
-					return ret
-				}
-				teamReviewers = append(teamReviewers, team)
-			} else {
-				user, ok := userReviewersMap[rID]
-				if !ok {
-					ctx.NotFound("", nil)
-					return ret
-				}
-				reviewers = append(reviewers, user)
-			}
-		}
-	}
-
-	ret.LabelIDs, ret.AssigneeIDs, ret.MilestoneID, ret.ProjectID = inputLabelIDs, inputAssigneeIDs, form.MilestoneID, form.ProjectID
-	ret.Reviewers, ret.TeamReviewers = reviewers, teamReviewers
-	return ret
-}
-
-// NewIssuePost response for creating new issue
-func NewIssuePost(ctx *context.Context) {
-	form := web.GetForm(ctx).(*forms.CreateIssueForm)
-	ctx.Data["Title"] = ctx.Tr("repo.issues.new")
-	ctx.Data["PageIsIssueList"] = true
-	ctx.Data["NewIssueChooseTemplate"] = issue_service.HasTemplatesOrContactLinks(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	ctx.Data["PullRequestWorkInProgressPrefixes"] = setting.Repository.PullRequest.WorkInProgressPrefixes
-	ctx.Data["IsAttachmentEnabled"] = setting.Attachment.Enabled
-	upload.AddUploadContext(ctx, "comment")
-
-	var (
-		repo        = ctx.Repo.Repository
-		attachments []string
-	)
-
-	validateRet := ValidateRepoMetasForNewIssue(ctx, *form, false)
-	if ctx.Written() {
-		return
-	}
-
-	labelIDs, assigneeIDs, milestoneID, projectID := validateRet.LabelIDs, validateRet.AssigneeIDs, validateRet.MilestoneID, validateRet.ProjectID
-
-	if projectID > 0 {
-		if !ctx.Repo.CanRead(unit.TypeProjects) {
-			// User must also be able to see the project.
-			ctx.Error(http.StatusBadRequest, "user hasn't permissions to read projects")
-			return
-		}
-	}
-
-	if setting.Attachment.Enabled {
-		attachments = form.Files
-	}
-
-	if ctx.HasError() {
-		ctx.JSONError(ctx.GetErrMsg())
-		return
-	}
-
-	if util.IsEmptyString(form.Title) {
-		ctx.JSONError(ctx.Tr("repo.issues.new.title_empty"))
-		return
-	}
-
-	content := form.Content
-	if filename := ctx.Req.Form.Get("template-file"); filename != "" {
-		if template, err := issue_template.UnmarshalFromRepo(ctx.Repo.GitRepo, ctx.Repo.Repository.DefaultBranch, filename); err == nil {
-			content = issue_template.RenderToMarkdown(template, ctx.Req.Form)
-		}
-	}
-
-	issue := &issues_model.Issue{
-		RepoID:      repo.ID,
-		Repo:        repo,
-		Title:       form.Title,
-		PosterID:    ctx.Doer.ID,
-		Poster:      ctx.Doer,
-		MilestoneID: milestoneID,
-		Content:     content,
-		Ref:         form.Ref,
-	}
-
-	if err := issue_service.NewIssue(ctx, repo, issue, labelIDs, attachments, assigneeIDs, projectID); err != nil {
-		if repo_model.IsErrUserDoesNotHaveAccessToRepo(err) {
-			ctx.Error(http.StatusBadRequest, "UserDoesNotHaveAccessToRepo", err.Error())
-		} else if errors.Is(err, user_model.ErrBlockedUser) {
-			ctx.JSONError(ctx.Tr("repo.issues.new.blocked_user"))
-		} else {
-			ctx.ServerError("NewIssue", err)
-		}
-		return
-	}
-
-	log.Trace("Issue created: %d/%d", repo.ID, issue.ID)
-	if ctx.FormString("redirect_after_creation") == "project" && projectID > 0 {
-		ctx.JSONRedirect(ctx.Repo.RepoLink + "/projects/" + strconv.FormatInt(projectID, 10))
-	} else {
-		ctx.JSONRedirect(issue.Link())
-	}
-}
-
-// roleDescriptor returns the role descriptor for a comment in/with the given repo, poster and issue
-func roleDescriptor(ctx stdCtx.Context, repo *repo_model.Repository, poster *user_model.User, issue *issues_model.Issue, hasOriginalAuthor bool) (issues_model.RoleDescriptor, error) {
-	roleDescriptor := issues_model.RoleDescriptor{}
-
-	if hasOriginalAuthor {
-		return roleDescriptor, nil
-	}
-
-	perm, err := access_model.GetUserRepoPermission(ctx, repo, poster)
-	if err != nil {
-		return roleDescriptor, err
-	}
-
-	// If the poster is the actual poster of the issue, enable Poster role.
-	roleDescriptor.IsPoster = issue.IsPoster(poster.ID)
-
-	// Check if the poster is owner of the repo.
-	if perm.IsOwner() {
-		// If the poster isn't an admin, enable the owner role.
-		if !poster.IsAdmin {
-			roleDescriptor.RoleInRepo = issues_model.RoleRepoOwner
-			return roleDescriptor, nil
-		}
-
-		// Otherwise check if poster is the real repo admin.
-		ok, err := access_model.IsUserRealRepoAdmin(ctx, repo, poster)
-		if err != nil {
-			return roleDescriptor, err
-		}
-		if ok {
-			roleDescriptor.RoleInRepo = issues_model.RoleRepoOwner
-			return roleDescriptor, nil
-		}
-	}
-
-	// If repo is organization, check Member role
-	if err := repo.LoadOwner(ctx); err != nil {
-		return roleDescriptor, err
-	}
-	if repo.Owner.IsOrganization() {
-		if isMember, err := organization.IsOrganizationMember(ctx, repo.Owner.ID, poster.ID); err != nil {
-			return roleDescriptor, err
-		} else if isMember {
-			roleDescriptor.RoleInRepo = issues_model.RoleRepoMember
-			return roleDescriptor, nil
-		}
-	}
-
-	// If the poster is the collaborator of the repo
-	if isCollaborator, err := repo_model.IsCollaborator(ctx, repo.ID, poster.ID); err != nil {
-		return roleDescriptor, err
-	} else if isCollaborator {
-		roleDescriptor.RoleInRepo = issues_model.RoleRepoCollaborator
-		return roleDescriptor, nil
-	}
-
-	hasMergedPR, err := issues_model.HasMergedPullRequestInRepo(ctx, repo.ID, poster.ID)
-	if err != nil {
-		return roleDescriptor, err
-	} else if hasMergedPR {
-		roleDescriptor.RoleInRepo = issues_model.RoleRepoContributor
-	} else if issue.IsPull {
-		// only display first time contributor in the first opening pull request
-		roleDescriptor.RoleInRepo = issues_model.RoleRepoFirstTimeContributor
-	}
-
-	return roleDescriptor, nil
-}
-
-func getBranchData(ctx *context.Context, issue *issues_model.Issue) {
-	ctx.Data["BaseBranch"] = nil
-	ctx.Data["HeadBranch"] = nil
-	ctx.Data["HeadUserName"] = nil
-	ctx.Data["BaseName"] = ctx.Repo.Repository.OwnerName
-	if issue.IsPull {
-		pull := issue.PullRequest
-		ctx.Data["BaseBranch"] = pull.BaseBranch
-		ctx.Data["HeadBranch"] = pull.HeadBranch
-		ctx.Data["HeadUserName"] = pull.MustHeadUserName(ctx)
-	}
-}
-
-// ViewIssue render issue view page
-func ViewIssue(ctx *context.Context) {
-	if ctx.PathParam(":type") == "issues" {
-		// If issue was requested we check if repo has external tracker and redirect
-		extIssueUnit, err := ctx.Repo.Repository.GetUnit(ctx, unit.TypeExternalTracker)
-		if err == nil && extIssueUnit != nil {
-			if extIssueUnit.ExternalTrackerConfig().ExternalTrackerStyle == markup.IssueNameStyleNumeric || extIssueUnit.ExternalTrackerConfig().ExternalTrackerStyle == "" {
-				metas := ctx.Repo.Repository.ComposeMetas(ctx)
-				metas["index"] = ctx.PathParam(":index")
-				res, err := vars.Expand(extIssueUnit.ExternalTrackerConfig().ExternalTrackerFormat, metas)
-				if err != nil {
-					log.Error("unable to expand template vars for issue url. issue: %s, err: %v", metas["index"], err)
-					ctx.ServerError("Expand", err)
-					return
-				}
-				ctx.Redirect(res)
-				return
-			}
-		} else if err != nil && !repo_model.IsErrUnitTypeNotExist(err) {
-			ctx.ServerError("GetUnit", err)
-			return
-		}
-	}
-
-	issue, err := issues_model.GetIssueByIndex(ctx, ctx.Repo.Repository.ID, ctx.PathParamInt64(":index"))
-	if err != nil {
-		if issues_model.IsErrIssueNotExist(err) {
-			ctx.NotFound("GetIssueByIndex", err)
-		} else {
-			ctx.ServerError("GetIssueByIndex", err)
-		}
-		return
-	}
-	if issue.Repo == nil {
-		issue.Repo = ctx.Repo.Repository
-	}
-
-	// Make sure type and URL matches.
-	if ctx.PathParam(":type") == "issues" && issue.IsPull {
-		ctx.Redirect(issue.Link())
-		return
-	} else if ctx.PathParam(":type") == "pulls" && !issue.IsPull {
-		ctx.Redirect(issue.Link())
-		return
-	}
-
-	if issue.IsPull {
-		MustAllowPulls(ctx)
-		if ctx.Written() {
-			return
-		}
-		ctx.Data["PageIsPullList"] = true
-		ctx.Data["PageIsPullConversation"] = true
-	} else {
-		MustEnableIssues(ctx)
-		if ctx.Written() {
-			return
-		}
-		ctx.Data["PageIsIssueList"] = true
-		ctx.Data["NewIssueChooseTemplate"] = issue_service.HasTemplatesOrContactLinks(ctx.Repo.Repository, ctx.Repo.GitRepo)
-	}
-
-	if issue.IsPull && !ctx.Repo.CanRead(unit.TypeIssues) {
-		ctx.Data["IssueDependencySearchType"] = "pulls"
-	} else if !issue.IsPull && !ctx.Repo.CanRead(unit.TypePullRequests) {
-		ctx.Data["IssueDependencySearchType"] = "issues"
-	} else {
-		ctx.Data["IssueDependencySearchType"] = "all"
-	}
-
-	ctx.Data["IsProjectsEnabled"] = ctx.Repo.CanRead(unit.TypeProjects)
-	ctx.Data["IsAttachmentEnabled"] = setting.Attachment.Enabled
-	upload.AddUploadContext(ctx, "comment")
-
-	if err = issue.LoadAttributes(ctx); err != nil {
-		ctx.ServerError("LoadAttributes", err)
-		return
-	}
-
-	if err = filterXRefComments(ctx, issue); err != nil {
-		ctx.ServerError("filterXRefComments", err)
-		return
-	}
-
-	ctx.Data["Title"] = fmt.Sprintf("#%d - %s", issue.Index, emoji.ReplaceAliases(issue.Title))
-
-	iw := new(issues_model.IssueWatch)
-	if ctx.Doer != nil {
-		iw.UserID = ctx.Doer.ID
-		iw.IssueID = issue.ID
-		iw.IsWatching, err = issues_model.CheckIssueWatch(ctx, ctx.Doer, issue)
-		if err != nil {
-			ctx.ServerError("CheckIssueWatch", err)
-			return
-		}
-	}
-	ctx.Data["IssueWatch"] = iw
-	issue.RenderedContent, err = markdown.RenderString(&markup.RenderContext{
-		Links: markup.Links{
-			Base: ctx.Repo.RepoLink,
-		},
-		Metas:   ctx.Repo.Repository.ComposeMetas(ctx),
-		GitRepo: ctx.Repo.GitRepo,
-		Repo:    ctx.Repo.Repository,
-		Ctx:     ctx,
-	}, issue.Content)
-	if err != nil {
-		ctx.ServerError("RenderString", err)
-		return
-	}
-
-	repo := ctx.Repo.Repository
-
-	// Get more information if it's a pull request.
-	if issue.IsPull {
-		if issue.PullRequest.HasMerged {
-			ctx.Data["DisableStatusChange"] = issue.PullRequest.HasMerged
-			PrepareMergedViewPullInfo(ctx, issue)
-		} else {
-			PrepareViewPullInfo(ctx, issue)
-			ctx.Data["DisableStatusChange"] = ctx.Data["IsPullRequestBroken"] == true && issue.IsClosed
-		}
-		if ctx.Written() {
-			return
-		}
-	}
-
-	pageMetaData := retrieveRepoIssueMetaData(ctx, repo, issue, issue.IsPull)
-	if ctx.Written() {
-		return
-	}
-	pageMetaData.LabelsData.SetSelectedLabels(issue.Labels)
-
-	if ctx.IsSigned {
-		// Update issue-user.
-		if err = activities_model.SetIssueReadBy(ctx, issue.ID, ctx.Doer.ID); err != nil {
-			ctx.ServerError("ReadBy", err)
-			return
-		}
-	}
-
-	var (
-		role                 issues_model.RoleDescriptor
-		ok                   bool
-		marked               = make(map[int64]issues_model.RoleDescriptor)
-		comment              *issues_model.Comment
-		participants         = make([]*user_model.User, 1, 10)
-		latestCloseCommentID int64
-	)
-	if ctx.Repo.Repository.IsTimetrackerEnabled(ctx) {
-		if ctx.IsSigned {
-			// Deal with the stopwatch
-			ctx.Data["IsStopwatchRunning"] = issues_model.StopwatchExists(ctx, ctx.Doer.ID, issue.ID)
-			if !ctx.Data["IsStopwatchRunning"].(bool) {
-				var exists bool
-				var swIssue *issues_model.Issue
-				if exists, _, swIssue, err = issues_model.HasUserStopwatch(ctx, ctx.Doer.ID); err != nil {
-					ctx.ServerError("HasUserStopwatch", err)
-					return
-				}
-				ctx.Data["HasUserStopwatch"] = exists
-				if exists {
-					// Add warning if the user has already a stopwatch
-					// Add link to the issue of the already running stopwatch
-					ctx.Data["OtherStopwatchURL"] = swIssue.Link()
-				}
-			}
-			ctx.Data["CanUseTimetracker"] = ctx.Repo.CanUseTimetracker(ctx, issue, ctx.Doer)
-		} else {
-			ctx.Data["CanUseTimetracker"] = false
-		}
-		if ctx.Data["WorkingUsers"], err = issues_model.TotalTimesForEachUser(ctx, &issues_model.FindTrackedTimesOptions{IssueID: issue.ID}); err != nil {
-			ctx.ServerError("TotalTimesForEachUser", err)
-			return
-		}
-	}
-
-	// Check if the user can use the dependencies
-	ctx.Data["CanCreateIssueDependencies"] = ctx.Repo.CanCreateIssueDependencies(ctx, ctx.Doer, issue.IsPull)
-
-	// check if dependencies can be created across repositories
-	ctx.Data["AllowCrossRepositoryDependencies"] = setting.Service.AllowCrossRepositoryDependencies
-
-	if issue.ShowRole, err = roleDescriptor(ctx, repo, issue.Poster, issue, issue.HasOriginalAuthor()); err != nil {
-		ctx.ServerError("roleDescriptor", err)
-		return
-	}
-	marked[issue.PosterID] = issue.ShowRole
-
-	// Render comments and fetch participants.
-	participants[0] = issue.Poster
-
-	if err := issue.Comments.LoadAttachmentsByIssue(ctx); err != nil {
-		ctx.ServerError("LoadAttachmentsByIssue", err)
-		return
-	}
-	if err := issue.Comments.LoadPosters(ctx); err != nil {
-		ctx.ServerError("LoadPosters", err)
-		return
-	}
-
-	for _, comment = range issue.Comments {
-		comment.Issue = issue
-
-		if comment.Type == issues_model.CommentTypeComment || comment.Type == issues_model.CommentTypeReview {
-			comment.RenderedContent, err = markdown.RenderString(&markup.RenderContext{
-				Links: markup.Links{
-					Base: ctx.Repo.RepoLink,
-				},
-				Metas:   ctx.Repo.Repository.ComposeMetas(ctx),
-				GitRepo: ctx.Repo.GitRepo,
-				Repo:    ctx.Repo.Repository,
-				Ctx:     ctx,
-			}, comment.Content)
-			if err != nil {
-				ctx.ServerError("RenderString", err)
-				return
-			}
-			// Check tag.
-			role, ok = marked[comment.PosterID]
-			if ok {
-				comment.ShowRole = role
-				continue
-			}
-
-			comment.ShowRole, err = roleDescriptor(ctx, repo, comment.Poster, issue, comment.HasOriginalAuthor())
-			if err != nil {
-				ctx.ServerError("roleDescriptor", err)
-				return
-			}
-			marked[comment.PosterID] = comment.ShowRole
-			participants = addParticipant(comment.Poster, participants)
-		} else if comment.Type == issues_model.CommentTypeLabel {
-			if err = comment.LoadLabel(ctx); err != nil {
-				ctx.ServerError("LoadLabel", err)
-				return
-			}
-		} else if comment.Type == issues_model.CommentTypeMilestone {
-			if err = comment.LoadMilestone(ctx); err != nil {
-				ctx.ServerError("LoadMilestone", err)
-				return
-			}
-			ghostMilestone := &issues_model.Milestone{
-				ID:   -1,
-				Name: ctx.Locale.TrString("repo.issues.deleted_milestone"),
-			}
-			if comment.OldMilestoneID > 0 && comment.OldMilestone == nil {
-				comment.OldMilestone = ghostMilestone
-			}
-			if comment.MilestoneID > 0 && comment.Milestone == nil {
-				comment.Milestone = ghostMilestone
-			}
-		} else if comment.Type == issues_model.CommentTypeProject {
-			if err = comment.LoadProject(ctx); err != nil {
-				ctx.ServerError("LoadProject", err)
-				return
-			}
-
-			ghostProject := &project_model.Project{
-				ID:    project_model.GhostProjectID,
-				Title: ctx.Locale.TrString("repo.issues.deleted_project"),
-			}
-
-			if comment.OldProjectID > 0 && comment.OldProject == nil {
-				comment.OldProject = ghostProject
-			}
-
-			if comment.ProjectID > 0 && comment.Project == nil {
-				comment.Project = ghostProject
-			}
-		} else if comment.Type == issues_model.CommentTypeProjectColumn {
-			if err = comment.LoadProject(ctx); err != nil {
-				ctx.ServerError("LoadProject", err)
-				return
-			}
-		} else if comment.Type == issues_model.CommentTypeAssignees || comment.Type == issues_model.CommentTypeReviewRequest {
-			if err = comment.LoadAssigneeUserAndTeam(ctx); err != nil {
-				ctx.ServerError("LoadAssigneeUserAndTeam", err)
-				return
-			}
-		} else if comment.Type == issues_model.CommentTypeRemoveDependency || comment.Type == issues_model.CommentTypeAddDependency {
-			if err = comment.LoadDepIssueDetails(ctx); err != nil {
-				if !issues_model.IsErrIssueNotExist(err) {
-					ctx.ServerError("LoadDepIssueDetails", err)
-					return
-				}
-			}
-		} else if comment.Type.HasContentSupport() {
-			comment.RenderedContent, err = markdown.RenderString(&markup.RenderContext{
-				Links: markup.Links{
-					Base: ctx.Repo.RepoLink,
-				},
-				Metas:   ctx.Repo.Repository.ComposeMetas(ctx),
-				GitRepo: ctx.Repo.GitRepo,
-				Repo:    ctx.Repo.Repository,
-				Ctx:     ctx,
-			}, comment.Content)
-			if err != nil {
-				ctx.ServerError("RenderString", err)
-				return
-			}
-			if err = comment.LoadReview(ctx); err != nil && !issues_model.IsErrReviewNotExist(err) {
-				ctx.ServerError("LoadReview", err)
-				return
-			}
-			participants = addParticipant(comment.Poster, participants)
-			if comment.Review == nil {
-				continue
-			}
-			if err = comment.Review.LoadAttributes(ctx); err != nil {
-				if !user_model.IsErrUserNotExist(err) {
-					ctx.ServerError("Review.LoadAttributes", err)
-					return
-				}
-				comment.Review.Reviewer = user_model.NewGhostUser()
-			}
-			if err = comment.Review.LoadCodeComments(ctx); err != nil {
-				ctx.ServerError("Review.LoadCodeComments", err)
-				return
-			}
-			for _, codeComments := range comment.Review.CodeComments {
-				for _, lineComments := range codeComments {
-					for _, c := range lineComments {
-						// Check tag.
-						role, ok = marked[c.PosterID]
-						if ok {
-							c.ShowRole = role
-							continue
-						}
-
-						c.ShowRole, err = roleDescriptor(ctx, repo, c.Poster, issue, c.HasOriginalAuthor())
-						if err != nil {
-							ctx.ServerError("roleDescriptor", err)
-							return
-						}
-						marked[c.PosterID] = c.ShowRole
-						participants = addParticipant(c.Poster, participants)
-					}
-				}
-			}
-			if err = comment.LoadResolveDoer(ctx); err != nil {
-				ctx.ServerError("LoadResolveDoer", err)
-				return
-			}
-		} else if comment.Type == issues_model.CommentTypePullRequestPush {
-			participants = addParticipant(comment.Poster, participants)
-			if err = comment.LoadPushCommits(ctx); err != nil {
-				ctx.ServerError("LoadPushCommits", err)
-				return
-			}
-			if !ctx.Repo.CanRead(unit.TypeActions) {
-				for _, commit := range comment.Commits {
-					commit.Status.HideActionsURL(ctx)
-					git_model.CommitStatusesHideActionsURL(ctx, commit.Statuses)
-				}
-			}
-		} else if comment.Type == issues_model.CommentTypeAddTimeManual ||
-			comment.Type == issues_model.CommentTypeStopTracking ||
-			comment.Type == issues_model.CommentTypeDeleteTimeManual {
-			// drop error since times could be pruned from DB..
-			_ = comment.LoadTime(ctx)
-			if comment.Content != "" {
-				// Content before v1.21 did store the formatted string instead of seconds,
-				// so "|" is used as delimiter to mark the new format
-				if comment.Content[0] != '|' {
-					// handle old time comments that have formatted text stored
-					comment.RenderedContent = templates.SanitizeHTML(comment.Content)
-					comment.Content = ""
-				} else {
-					// else it's just a duration in seconds to pass on to the frontend
-					comment.Content = comment.Content[1:]
-				}
-			}
-		}
-
-		if comment.Type == issues_model.CommentTypeClose || comment.Type == issues_model.CommentTypeMergePull {
-			// record ID of the latest closed/merged comment.
-			// if PR is closed, the comments whose type is CommentTypePullRequestPush(29) after latestCloseCommentID won't be rendered.
-			latestCloseCommentID = comment.ID
-		}
-	}
-
-	ctx.Data["LatestCloseCommentID"] = latestCloseCommentID
-
-	// Combine multiple label assignments into a single comment
-	combineLabelComments(issue)
-
-	getBranchData(ctx, issue)
-	if issue.IsPull {
-		pull := issue.PullRequest
-		pull.Issue = issue
-		canDelete := false
-		allowMerge := false
-		canWriteToHeadRepo := false
-
-		if ctx.IsSigned {
-			if err := pull.LoadHeadRepo(ctx); err != nil {
-				log.Error("LoadHeadRepo: %v", err)
-			} else if pull.HeadRepo != nil {
-				perm, err := access_model.GetUserRepoPermission(ctx, pull.HeadRepo, ctx.Doer)
-				if err != nil {
-					ctx.ServerError("GetUserRepoPermission", err)
-					return
-				}
-				if perm.CanWrite(unit.TypeCode) {
-					// Check if branch is not protected
-					if pull.HeadBranch != pull.HeadRepo.DefaultBranch {
-						if protected, err := git_model.IsBranchProtected(ctx, pull.HeadRepo.ID, pull.HeadBranch); err != nil {
-							log.Error("IsProtectedBranch: %v", err)
-						} else if !protected {
-							canDelete = true
-							ctx.Data["DeleteBranchLink"] = issue.Link() + "/cleanup"
-						}
-					}
-					canWriteToHeadRepo = true
-				}
-			}
-
-			if err := pull.LoadBaseRepo(ctx); err != nil {
-				log.Error("LoadBaseRepo: %v", err)
-			}
-			perm, err := access_model.GetUserRepoPermission(ctx, pull.BaseRepo, ctx.Doer)
-			if err != nil {
-				ctx.ServerError("GetUserRepoPermission", err)
-				return
-			}
-			if !canWriteToHeadRepo { // maintainers maybe allowed to push to head repo even if they can't write to it
-				canWriteToHeadRepo = pull.AllowMaintainerEdit && perm.CanWrite(unit.TypeCode)
-			}
-			allowMerge, err = pull_service.IsUserAllowedToMerge(ctx, pull, perm, ctx.Doer)
-			if err != nil {
-				ctx.ServerError("IsUserAllowedToMerge", err)
-				return
-			}
-
-			if ctx.Data["CanMarkConversation"], err = issues_model.CanMarkConversation(ctx, issue, ctx.Doer); err != nil {
-				ctx.ServerError("CanMarkConversation", err)
-				return
-			}
-		}
-
-		ctx.Data["CanWriteToHeadRepo"] = canWriteToHeadRepo
-		ctx.Data["ShowMergeInstructions"] = canWriteToHeadRepo
-		ctx.Data["AllowMerge"] = allowMerge
-
-		prUnit, err := repo.GetUnit(ctx, unit.TypePullRequests)
-		if err != nil {
-			ctx.ServerError("GetUnit", err)
-			return
-		}
-		prConfig := prUnit.PullRequestsConfig()
-
-		ctx.Data["AutodetectManualMerge"] = prConfig.AutodetectManualMerge
-
-		var mergeStyle repo_model.MergeStyle
-		// Check correct values and select default
-		if ms, ok := ctx.Data["MergeStyle"].(repo_model.MergeStyle); !ok ||
-			!prConfig.IsMergeStyleAllowed(ms) {
-			defaultMergeStyle := prConfig.GetDefaultMergeStyle()
-			if prConfig.IsMergeStyleAllowed(defaultMergeStyle) && !ok {
-				mergeStyle = defaultMergeStyle
-			} else if prConfig.AllowMerge {
-				mergeStyle = repo_model.MergeStyleMerge
-			} else if prConfig.AllowRebase {
-				mergeStyle = repo_model.MergeStyleRebase
-			} else if prConfig.AllowRebaseMerge {
-				mergeStyle = repo_model.MergeStyleRebaseMerge
-			} else if prConfig.AllowSquash {
-				mergeStyle = repo_model.MergeStyleSquash
-			} else if prConfig.AllowFastForwardOnly {
-				mergeStyle = repo_model.MergeStyleFastForwardOnly
-			} else if prConfig.AllowManualMerge {
-				mergeStyle = repo_model.MergeStyleManuallyMerged
-			}
-		}
-
-		ctx.Data["MergeStyle"] = mergeStyle
-
-		defaultMergeMessage, defaultMergeBody, err := pull_service.GetDefaultMergeMessage(ctx, ctx.Repo.GitRepo, pull, mergeStyle)
-		if err != nil {
-			ctx.ServerError("GetDefaultMergeMessage", err)
-			return
-		}
-		ctx.Data["DefaultMergeMessage"] = defaultMergeMessage
-		ctx.Data["DefaultMergeBody"] = defaultMergeBody
-
-		defaultSquashMergeMessage, defaultSquashMergeBody, err := pull_service.GetDefaultMergeMessage(ctx, ctx.Repo.GitRepo, pull, repo_model.MergeStyleSquash)
-		if err != nil {
-			ctx.ServerError("GetDefaultSquashMergeMessage", err)
-			return
-		}
-		ctx.Data["DefaultSquashMergeMessage"] = defaultSquashMergeMessage
-		ctx.Data["DefaultSquashMergeBody"] = defaultSquashMergeBody
-
-		pb, err := git_model.GetFirstMatchProtectedBranchRule(ctx, pull.BaseRepoID, pull.BaseBranch)
-		if err != nil {
-			ctx.ServerError("LoadProtectedBranch", err)
-			return
-		}
-
-		if pb != nil {
-			pb.Repo = pull.BaseRepo
-			ctx.Data["ProtectedBranch"] = pb
-			ctx.Data["IsBlockedByApprovals"] = !issues_model.HasEnoughApprovals(ctx, pb, pull)
-			ctx.Data["IsBlockedByRejection"] = issues_model.MergeBlockedByRejectedReview(ctx, pb, pull)
-			ctx.Data["IsBlockedByOfficialReviewRequests"] = issues_model.MergeBlockedByOfficialReviewRequests(ctx, pb, pull)
-			ctx.Data["IsBlockedByOutdatedBranch"] = issues_model.MergeBlockedByOutdatedBranch(pb, pull)
-			ctx.Data["GrantedApprovals"] = issues_model.GetGrantedApprovalsCount(ctx, pb, pull)
-			ctx.Data["RequireSigned"] = pb.RequireSignedCommits
-			ctx.Data["ChangedProtectedFiles"] = pull.ChangedProtectedFiles
-			ctx.Data["IsBlockedByChangedProtectedFiles"] = len(pull.ChangedProtectedFiles) != 0
-			ctx.Data["ChangedProtectedFilesNum"] = len(pull.ChangedProtectedFiles)
-			ctx.Data["RequireApprovalsWhitelist"] = pb.EnableApprovalsWhitelist
-		}
-		ctx.Data["WillSign"] = false
-		if ctx.Doer != nil {
-			sign, key, _, err := asymkey_service.SignMerge(ctx, pull, ctx.Doer, pull.BaseRepo.RepoPath(), pull.BaseBranch, pull.GetGitRefName())
-			ctx.Data["WillSign"] = sign
-			ctx.Data["SigningKey"] = key
-			if err != nil {
-				if asymkey_service.IsErrWontSign(err) {
-					ctx.Data["WontSignReason"] = err.(*asymkey_service.ErrWontSign).Reason
-				} else {
-					ctx.Data["WontSignReason"] = "error"
-					log.Error("Error whilst checking if could sign pr %d in repo %s. Error: %v", pull.ID, pull.BaseRepo.FullName(), err)
-				}
-			}
-		} else {
-			ctx.Data["WontSignReason"] = "not_signed_in"
-		}
-
-		isPullBranchDeletable := canDelete &&
-			pull.HeadRepo != nil &&
-			git.IsBranchExist(ctx, pull.HeadRepo.RepoPath(), pull.HeadBranch) &&
-			(!pull.HasMerged || ctx.Data["HeadBranchCommitID"] == ctx.Data["PullHeadCommitID"])
-
-		if isPullBranchDeletable && pull.HasMerged {
-			exist, err := issues_model.HasUnmergedPullRequestsByHeadInfo(ctx, pull.HeadRepoID, pull.HeadBranch)
-			if err != nil {
-				ctx.ServerError("HasUnmergedPullRequestsByHeadInfo", err)
-				return
-			}
-
-			isPullBranchDeletable = !exist
-		}
-		ctx.Data["IsPullBranchDeletable"] = isPullBranchDeletable
-
-		stillCanManualMerge := func() bool {
-			if pull.HasMerged || issue.IsClosed || !ctx.IsSigned {
-				return false
-			}
-			if pull.CanAutoMerge() || pull.IsWorkInProgress(ctx) || pull.IsChecking() {
-				return false
-			}
-			if allowMerge && prConfig.AllowManualMerge {
-				return true
-			}
-
-			return false
-		}
-
-		ctx.Data["StillCanManualMerge"] = stillCanManualMerge()
-
-		// Check if there is a pending pr merge
-		ctx.Data["HasPendingPullRequestMerge"], ctx.Data["PendingPullRequestMerge"], err = pull_model.GetScheduledMergeByPullID(ctx, pull.ID)
-		if err != nil {
-			ctx.ServerError("GetScheduledMergeByPullID", err)
-			return
-		}
-	}
-
-	// Get Dependencies
-	blockedBy, err := issue.BlockedByDependencies(ctx, db.ListOptions{})
-	if err != nil {
-		ctx.ServerError("BlockedByDependencies", err)
-		return
-	}
-	ctx.Data["BlockedByDependencies"], ctx.Data["BlockedByDependenciesNotPermitted"] = checkBlockedByIssues(ctx, blockedBy)
-	if ctx.Written() {
-		return
-	}
-
-	blocking, err := issue.BlockingDependencies(ctx)
-	if err != nil {
-		ctx.ServerError("BlockingDependencies", err)
-		return
-	}
-
-	ctx.Data["BlockingDependencies"], ctx.Data["BlockingDependenciesNotPermitted"] = checkBlockedByIssues(ctx, blocking)
-	if ctx.Written() {
-		return
-	}
-
-	var pinAllowed bool
-	if !issue.IsPinned() {
-		pinAllowed, err = issues_model.IsNewPinAllowed(ctx, issue.RepoID, issue.IsPull)
-		if err != nil {
-			ctx.ServerError("IsNewPinAllowed", err)
-			return
-		}
-	} else {
-		pinAllowed = true
-	}
-
-	ctx.Data["Participants"] = participants
-	ctx.Data["NumParticipants"] = len(participants)
-	ctx.Data["Issue"] = issue
-	ctx.Data["Reference"] = issue.Ref
-	ctx.Data["SignInLink"] = setting.AppSubURL + "/user/login?redirect_to=" + url.QueryEscape(ctx.Data["Link"].(string))
-	ctx.Data["IsIssuePoster"] = ctx.IsSigned && issue.IsPoster(ctx.Doer.ID)
-	ctx.Data["HasIssuesOrPullsWritePermission"] = ctx.Repo.CanWriteIssuesOrPulls(issue.IsPull)
-	ctx.Data["HasProjectsWritePermission"] = ctx.Repo.CanWrite(unit.TypeProjects)
-	ctx.Data["IsRepoAdmin"] = ctx.IsSigned && (ctx.Repo.IsAdmin() || ctx.Doer.IsAdmin)
-	ctx.Data["LockReasons"] = setting.Repository.Issue.LockReasons
-	ctx.Data["RefEndName"] = git.RefName(issue.Ref).ShortName()
-	ctx.Data["NewPinAllowed"] = pinAllowed
-	ctx.Data["PinEnabled"] = setting.Repository.Issue.MaxPinned != 0
-
-	var hiddenCommentTypes *big.Int
-	if ctx.IsSigned {
-		val, err := user_model.GetUserSetting(ctx, ctx.Doer.ID, user_model.SettingsKeyHiddenCommentTypes)
-		if err != nil {
-			ctx.ServerError("GetUserSetting", err)
-			return
-		}
-		hiddenCommentTypes, _ = new(big.Int).SetString(val, 10) // we can safely ignore the failed conversion here
-	}
-	ctx.Data["ShouldShowCommentType"] = func(commentType issues_model.CommentType) bool {
-		return hiddenCommentTypes == nil || hiddenCommentTypes.Bit(int(commentType)) == 0
-	}
-	// For sidebar
-	PrepareBranchList(ctx)
-
-	if ctx.Written() {
-		return
-	}
-
-	tags, err := repo_model.GetTagNamesByRepoID(ctx, ctx.Repo.Repository.ID)
-	if err != nil {
-		ctx.ServerError("GetTagNamesByRepoID", err)
-		return
-	}
-	ctx.Data["Tags"] = tags
-
-	ctx.Data["CanBlockUser"] = func(blocker, blockee *user_model.User) bool {
-		return user_service.CanBlockUser(ctx, ctx.Doer, blocker, blockee)
-	}
-
-	ctx.HTML(http.StatusOK, tplIssueView)
-}
-
-// checkBlockedByIssues return canRead and notPermitted
-func checkBlockedByIssues(ctx *context.Context, blockers []*issues_model.DependencyInfo) (canRead, notPermitted []*issues_model.DependencyInfo) {
-	repoPerms := make(map[int64]access_model.Permission)
-	repoPerms[ctx.Repo.Repository.ID] = ctx.Repo.Permission
-	for _, blocker := range blockers {
-		// Get the permissions for this repository
-		// If the repo ID exists in the map, return the exist permissions
-		// else get the permission and add it to the map
-		var perm access_model.Permission
-		existPerm, ok := repoPerms[blocker.RepoID]
-		if ok {
-			perm = existPerm
-		} else {
-			var err error
-			perm, err = access_model.GetUserRepoPermission(ctx, &blocker.Repository, ctx.Doer)
-			if err != nil {
-				ctx.ServerError("GetUserRepoPermission", err)
-				return nil, nil
-			}
-			repoPerms[blocker.RepoID] = perm
-		}
-		if perm.CanReadIssuesOrPulls(blocker.Issue.IsPull) {
-			canRead = append(canRead, blocker)
-		} else {
-			notPermitted = append(notPermitted, blocker)
-		}
-	}
-	sortDependencyInfo(canRead)
-	sortDependencyInfo(notPermitted)
-	return canRead, notPermitted
-}
-
-func sortDependencyInfo(blockers []*issues_model.DependencyInfo) {
-	sort.Slice(blockers, func(i, j int) bool {
-		if blockers[i].RepoID == blockers[j].RepoID {
-			return blockers[i].Issue.CreatedUnix < blockers[j].Issue.CreatedUnix
-		}
-		return blockers[i].RepoID < blockers[j].RepoID
-	})
 }
 
 // GetActionIssue will return the issue which is used in the context.
@@ -2578,40 +575,6 @@ func ChangeIssueReaction(ctx *context.Context) {
 	})
 }
 
-func addParticipant(poster *user_model.User, participants []*user_model.User) []*user_model.User {
-	for _, part := range participants {
-		if poster.ID == part.ID {
-			return participants
-		}
-	}
-	return append(participants, poster)
-}
-
-func filterXRefComments(ctx *context.Context, issue *issues_model.Issue) error {
-	// Remove comments that the user has no permissions to see
-	for i := 0; i < len(issue.Comments); {
-		c := issue.Comments[i]
-		if issues_model.CommentTypeIsRef(c.Type) && c.RefRepoID != issue.RepoID && c.RefRepoID != 0 {
-			var err error
-			// Set RefRepo for description in template
-			c.RefRepo, err = repo_model.GetRepositoryByID(ctx, c.RefRepoID)
-			if err != nil {
-				return err
-			}
-			perm, err := access_model.GetUserRepoPermission(ctx, c.RefRepo, ctx.Doer)
-			if err != nil {
-				return err
-			}
-			if !perm.CanReadIssuesOrPulls(c.RefIsPull) {
-				issue.Comments = append(issue.Comments[:i], issue.Comments[i+1:]...)
-				continue
-			}
-		}
-		i++
-	}
-	return nil
-}
-
 // GetIssueAttachments returns attachments for the issue
 func GetIssueAttachments(ctx *context.Context) {
 	issue := GetActionIssue(ctx)
@@ -2681,73 +644,6 @@ func attachmentsHTML(ctx *context.Context, attachments []*repo_model.Attachment,
 	return attachHTML
 }
 
-// combineLabelComments combine the nearby label comments as one.
-func combineLabelComments(issue *issues_model.Issue) {
-	var prev, cur *issues_model.Comment
-	for i := 0; i < len(issue.Comments); i++ {
-		cur = issue.Comments[i]
-		if i > 0 {
-			prev = issue.Comments[i-1]
-		}
-		if i == 0 || cur.Type != issues_model.CommentTypeLabel ||
-			(prev != nil && prev.PosterID != cur.PosterID) ||
-			(prev != nil && cur.CreatedUnix-prev.CreatedUnix >= 60) {
-			if cur.Type == issues_model.CommentTypeLabel && cur.Label != nil {
-				if cur.Content != "1" {
-					cur.RemovedLabels = append(cur.RemovedLabels, cur.Label)
-				} else {
-					cur.AddedLabels = append(cur.AddedLabels, cur.Label)
-				}
-			}
-			continue
-		}
-
-		if cur.Label != nil { // now cur MUST be label comment
-			if prev.Type == issues_model.CommentTypeLabel { // we can combine them only prev is a label comment
-				if cur.Content != "1" {
-					// remove labels from the AddedLabels list if the label that was removed is already
-					// in this list, and if it's not in this list, add the label to RemovedLabels
-					addedAndRemoved := false
-					for i, label := range prev.AddedLabels {
-						if cur.Label.ID == label.ID {
-							prev.AddedLabels = append(prev.AddedLabels[:i], prev.AddedLabels[i+1:]...)
-							addedAndRemoved = true
-							break
-						}
-					}
-					if !addedAndRemoved {
-						prev.RemovedLabels = append(prev.RemovedLabels, cur.Label)
-					}
-				} else {
-					// remove labels from the RemovedLabels list if the label that was added is already
-					// in this list, and if it's not in this list, add the label to AddedLabels
-					removedAndAdded := false
-					for i, label := range prev.RemovedLabels {
-						if cur.Label.ID == label.ID {
-							prev.RemovedLabels = append(prev.RemovedLabels[:i], prev.RemovedLabels[i+1:]...)
-							removedAndAdded = true
-							break
-						}
-					}
-					if !removedAndAdded {
-						prev.AddedLabels = append(prev.AddedLabels, cur.Label)
-					}
-				}
-				prev.CreatedUnix = cur.CreatedUnix
-				// remove the current comment since it has been combined to prev comment
-				issue.Comments = append(issue.Comments[:i], issue.Comments[i+1:]...)
-				i--
-			} else { // if prev is not a label comment, start a new group
-				if cur.Content != "1" {
-					cur.RemovedLabels = append(cur.RemovedLabels, cur.Label)
-				} else {
-					cur.AddedLabels = append(cur.AddedLabels, cur.Label)
-				}
-			}
-		}
-	}
-}
-
 // get all teams that current user can mention
 func handleTeamMentions(ctx *context.Context) {
 	if ctx.Doer == nil || !ctx.Repo.Owner.IsOrganization() {
@@ -2786,54 +682,4 @@ func handleTeamMentions(ctx *context.Context) {
 	ctx.Data["MentionableTeams"] = teams
 	ctx.Data["MentionableTeamsOrg"] = ctx.Repo.Owner.Name
 	ctx.Data["MentionableTeamsOrgAvatar"] = ctx.Repo.Owner.AvatarLink(ctx)
-}
-
-type userSearchInfo struct {
-	UserID     int64  `json:"user_id"`
-	UserName   string `json:"username"`
-	AvatarLink string `json:"avatar_link"`
-	FullName   string `json:"full_name"`
-}
-
-type userSearchResponse struct {
-	Results []*userSearchInfo `json:"results"`
-}
-
-// IssuePosters get posters for current repo's issues/pull requests
-func IssuePosters(ctx *context.Context) {
-	issuePosters(ctx, false)
-}
-
-func PullPosters(ctx *context.Context) {
-	issuePosters(ctx, true)
-}
-
-func issuePosters(ctx *context.Context, isPullList bool) {
-	repo := ctx.Repo.Repository
-	search := strings.TrimSpace(ctx.FormString("q"))
-	posters, err := repo_model.GetIssuePostersWithSearch(ctx, repo, isPullList, search, setting.UI.DefaultShowFullName)
-	if err != nil {
-		ctx.JSON(http.StatusInternalServerError, err)
-		return
-	}
-
-	if search == "" && ctx.Doer != nil {
-		// the returned posters slice only contains limited number of users,
-		// to make the current user (doer) can quickly filter their own issues, always add doer to the posters slice
-		if !slices.ContainsFunc(posters, func(user *user_model.User) bool { return user.ID == ctx.Doer.ID }) {
-			posters = append(posters, ctx.Doer)
-		}
-	}
-
-	posters = shared_user.MakeSelfOnTop(ctx.Doer, posters)
-
-	resp := &userSearchResponse{}
-	resp.Results = make([]*userSearchInfo, len(posters))
-	for i, user := range posters {
-		resp.Results[i] = &userSearchInfo{UserID: user.ID, UserName: user.Name, AvatarLink: user.AvatarLink(ctx)}
-		if setting.UI.DefaultShowFullName {
-			resp.Results[i].FullName = user.FullName
-		}
-	}
-	ctx.JSON(http.StatusOK, resp)
 }
